{{cookiecutter.nombre_proyecto}}/requirements.txt
altgraph==0.17.4
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
greenlet==3.1.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pefile==2024.8.26
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
pyzmq==26.2.0
requests==2.32.3
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.35
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
